requirements.txt
alembic==1.10.3
attrs==23.1.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
distro==1.8.0
docker==6.0.1
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
Flask==2.1.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.3
idna==3.4
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==3.2.0
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.1
mysql-connector-python==8.0.33
packaging==23.1
paramiko==3.1.0
protobuf==3.20.3
pycparser==2.21
Pygments==2.15.1
PyMySQL==1.0.3
PyNaCl==1.5.0
pyrsistent==0.19.3
python-dotenv==0.21.1
PyYAML==5.4.1
requests==2.28.2
six==1.16.0
SQLAlchemy==1.4.41
texttable==1.6.7
typing_extensions==4.5.0
urllib3==1.26.15
websocket-client==0.59.0
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
